Split the e2e sources out of src
Move the e2e test to e2e/ and the shared codegen bindings to
artifacts/ (regenerated so the artifact import paths match the new
location), making room for a sibling deployment/ folder.
