+++ packages/aztec/scripts/.gitignore
@@ -1,3 +1,5 @@
 node_modules/
 # `aztec codegen` scratch file, written to the cwd it is invoked from
 codegenCache.json
+# Un-ignore the codegen bindings: the root .gitignore ignores artifacts/ for Foundry.
+!artifacts/
