+++ packages/aztec/scripts/.prettierignore
@@ -1,2 +1,2 @@
 # Generated by `aztec codegen`; keep the generator's own formatting.
-src/artifacts/
+artifacts/
