+++ packages/aztec/scripts/e2e/circuit-version.ts
