+++ packages/aztec/scripts/e2e/run-e2e.ts
@@ -18,8 +18,8 @@ import { createAztecNodeDebugClient } from "@aztec/stdlib/interfaces/client"
 import { EmbeddedWallet } from "@aztec/wallets/embedded"
 import { registerInitialLocalNetworkAccountsInWallet } from "@aztec/wallets/testing"
 import { readFileSync } from "fs"
-import { AgeGateContract } from "./artifacts/AgeGate.js"
-import { ZKPassportRegistryContract } from "./artifacts/ZKPassportRegistry.js"
+import { AgeGateContract } from "../artifacts/AgeGate.js"
+import { ZKPassportRegistryContract } from "../artifacts/ZKPassportRegistry.js"
 import { packCircuitManifestVersion } from "./circuit-version.js"
 
 const NODE_URL = process.env.AZTEC_NODE_URL ?? "http://localhost:8080"
